\name{map_png}
\alias{map_png}
\title{Plotting image with ggmap
\code{map_png} uses ggmap to plot a map image}
\usage{
  map_png(map_info, ...)
}
\arguments{
  \item{map_info}{named vector contains the info for the
  map: center: c(lat,lng) filename: string size: c(w,h)
  zoom: int}

  \item{...}{arguments for ggmap}
}
\value{
  a ggmap object
}
\description{
  Plotting image with ggmap \code{map_png} uses ggmap to
  plot a map image
}
\seealso{
  \code{\link{ggmap}}
}

