\name{readmaptile}
\alias{readmaptile}
\title{Read the PNG and convert to raster ggmap}
\usage{
  readmaptile(map_info)
}
\arguments{
  \item{map_info}{a list of meta information
  (filename,center,zoom,size)}
}
\value{
  ggmap class raster_map
}
\description{
  Read the PNG and convert to raster ggmap
}
\seealso{
  \code{\link{ggmap}}
}

